Move inputworker to communication
